student_welfare_service/requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
mysqlclient==2.2.0
Flask-WTF==1.1.1
python-dotenv